Renames MAPPER field to INSTANCE
According to the convention explained in chapter 4.1 the static field declared in the mapper interface should be named INSTANCE.
However, my proposal is to remove this line completely here as this example appears in chapter 3.5. I myself asked my why this mapper instance has to be declared. Same with the `@InheritInverseConfiguration` annotation. You use it here but it has not yet been introduced 
here, so I would revert this to a conventional mapping annotation.