Mapper, Mapping, Mappings - Change RetentionPolicy to RUNTIME
Changing for Mapper, Mapping, Mappings the RetentionPolicy from CLASS to RUNTIME. This is to get at runtime the defintion of the mapping. Can be used for example to convert a sorting definition of a database object that is used in a query in the corresponding sorting defintion of the dto that can be used in the frontend.